ccl: move auditloggingccl to sql/auditlogging
This is part of an ongoing cleanup for pkg/ccl.

Epic: none
Release note: None
