--- pkg/ccl/auditloggingccl/BUILD.bazel
@@ -1,44 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "auditloggingccl",
-    srcs = ["audit_log_config.go"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/auditloggingccl",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/ccl/utilccl",
-        "//pkg/settings",
-        "//pkg/settings/cluster",
-        "//pkg/sql/auditlogging",
-        "//pkg/util/log",
-        "@com_github_cockroachdb_errors//:errors",
-    ],
-)
-
-go_test(
-    name = "auditloggingccl_test",
-    srcs = [
-        "audit_logging_test.go",
-        "main_test.go",
-    ],
-    embed = [":auditloggingccl"],
-    deps = [
-        "//pkg/base",
-        "//pkg/ccl",
-        "//pkg/security/securityassets",
-        "//pkg/security/securitytest",
-        "//pkg/security/username",
-        "//pkg/server",
-        "//pkg/testutils",
-        "//pkg/testutils/serverutils",
-        "//pkg/testutils/sqlutils",
-        "//pkg/testutils/testcluster",
-        "//pkg/util/leaktest",
-        "//pkg/util/log",
-        "//pkg/util/log/logpb",
-        "//pkg/util/log/logtestutils",
-        "//pkg/util/randutil",
-        "@com_github_cockroachdb_errors//:errors",
-        "@com_github_stretchr_testify//require",
-    ],
-)
