+++ pkg/sql/auditlogging/BUILD.bazel
@@ -17,6 +17,7 @@ go_library(
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/sem/tree",
+        "//pkg/util/log",
         "//pkg/util/log/eventpb",
         "//pkg/util/log/logpb",
         "//pkg/util/syncutil",
@@ -27,13 +28,33 @@ go_library(
 
 go_test(
     name = "auditlogging_test",
-    srcs = ["audit_log_test.go"],
+    srcs = [
+        "audit_log_test.go",
+        "audit_logging_test.go",
+        "main_test.go",
+    ],
     data = glob(["testdata/**"]),
     embed = [":auditlogging"],
     deps = [
+        "//pkg/base",
+        "//pkg/security/securityassets",
+        "//pkg/security/securitytest",
+        "//pkg/security/username",
+        "//pkg/server",
         "//pkg/settings/rulebasedscanner",
+        "//pkg/testutils",
         "//pkg/testutils/datapathutils",
+        "//pkg/testutils/serverutils",
+        "//pkg/testutils/sqlutils",
+        "//pkg/testutils/testcluster",
+        "//pkg/util/leaktest",
+        "//pkg/util/log",
+        "//pkg/util/log/logpb",
+        "//pkg/util/log/logtestutils",
+        "//pkg/util/randutil",
         "@com_github_cockroachdb_datadriven//:datadriven",
+        "@com_github_cockroachdb_errors//:errors",
         "@com_github_kr_pretty//:pretty",
+        "@com_github_stretchr_testify//require",
     ],
 )
