+++ pkg/sql/auditlogging/audit_logging_test.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package auditloggingccl
+package auditlogging_test
 
 import (
 	"context"
